Move DebugClientLogger into swatch-common-resteasy-client
