+++ swatch-contracts/src/main/resources/application.properties
@@ -168,12 +168,12 @@ quarkus.rest-client."com.redhat.swatch.clients.rh.partner.gateway.api.resources.
 quarkus.rest-client."com.redhat.swatch.clients.rh.partner.gateway.api.resources.PartnerApi".trust-store=${TRUSTSTORE_RESOURCE:}
 quarkus.rest-client."com.redhat.swatch.clients.rh.partner.gateway.api.resources.PartnerApi".trust-store-password=${TRUSTSTORE_PASSWORD:}
 quarkus.rest-client."com.redhat.swatch.clients.rh.partner.gateway.api.resources.PartnerApi".scope=jakarta.enterprise.context.ApplicationScoped
-quarkus.rest-client."com.redhat.swatch.clients.rh.partner.gateway.api.resources.PartnerApi".providers=com.redhat.swatch.contract.config.DebugClientLogger
+quarkus.rest-client."com.redhat.swatch.clients.rh.partner.gateway.api.resources.PartnerApi".providers=com.redhat.swatch.resteasy.client.DebugClientLogger
 
 # rest-client debug logging
 # NOTE: all of uri-filter-regex, log category level, and logging.scope need to be set for a request/response to be logged.
 rest-client-debug-logging.uri-filter=.*(partnerSubscriptions|search|products|export).*
-quarkus.log.category."com.redhat.swatch.contract.config.DebugClientLogger".level=DEBUG
+quarkus.log.category."com.redhat.swatch.resteasy.client.DebugClientLogger".level=DEBUG
 %test.quarkus.rest-client.logging.scope=request-response
 %dev.quarkus.rest-client.logging.scope=request-response
 %ephemeral.quarkus.rest-client.logging.scope=request-response
@@ -186,7 +186,7 @@ quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.Search
 quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.SearchApi".trust-store-password=${TRUSTSTORE_PASSWORD:}
 quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.SearchApi".scope=jakarta.enterprise.context.ApplicationScoped
 quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.SearchApi".connection-pool-size=${SUBSCRIPTION_MAX_CONNECTIONS}
-quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.SearchApi".providers=com.redhat.swatch.contract.config.DebugClientLogger
+quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.SearchApi".providers=com.redhat.swatch.resteasy.client.DebugClientLogger
 # Setting up the timeout to 2 minutes instead of 30 seconds (default)
 quarkus.rest-client."com.redhat.swatch.clients.subscription.api.resources.SearchApi".read-timeout=120000
 
@@ -198,7 +198,7 @@ quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi"
 quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi".trust-store=${TRUSTSTORE_RESOURCE:}
 quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi".trust-store-password=${TRUSTSTORE_PASSWORD:}
 quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi".scope=jakarta.enterprise.context.ApplicationScoped
-quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi".providers=com.redhat.swatch.contract.config.DebugClientLogger
+quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi".providers=com.redhat.swatch.resteasy.client.DebugClientLogger
 # Setting up the timeout to 60 seconds instead of 30 seconds (default)
 quarkus.rest-client."com.redhat.swatch.clients.product.api.resources.ProductApi".read-timeout=90000
