+++ swatch-metrics/src/main/resources/application.properties
@@ -91,10 +91,19 @@ quarkus.log.category."io.quarkus.http.access-log".handlers=ACCESS_LOG
 quarkus.log.category."io.quarkus.http.access-log".use-parent-handlers=false
 quarkus.log.category."org.jboss.resteasy.reactive.common.core.AbstractResteasyReactiveContext".level=DEBUG
 
+# rest-client debug logging
+# NOTE: all of uri-filter-regex, log category level, and logging.scope need to be set for a request/response to be logged.
+quarkus.log.category."com.redhat.swatch.resteasy.client.DebugClientLogger".level=DEBUG
+%test.quarkus.rest-client.logging.scope=request-response
+%dev.quarkus.rest-client.logging.scope=request-response
+%ephemeral.quarkus.rest-client.logging.scope=request-response
+
 quarkus.rest-client."com.redhat.swatch.clients.prometheus.api.resources.QueryApi".url=${rhsm-subscriptions.metering.prometheus.client.url}
 quarkus.rest-client."com.redhat.swatch.clients.prometheus.api.resources.QueryApi".scope=jakarta.enterprise.context.ApplicationScoped
+quarkus.rest-client."com.redhat.swatch.clients.prometheus.api.resources.QueryApi".providers=com.redhat.swatch.resteasy.client.DebugClientLogger
 quarkus.rest-client."com.redhat.swatch.clients.prometheus.api.resources.QueryRangeApi".url=${rhsm-subscriptions.metering.prometheus.client.url}
 quarkus.rest-client."com.redhat.swatch.clients.prometheus.api.resources.QueryRangeApi".scope=jakarta.enterprise.context.ApplicationScoped
+quarkus.rest-client."com.redhat.swatch.clients.prometheus.api.resources.QueryRangeApi".providers=com.redhat.swatch.resteasy.client.DebugClientLogger
 
 # Kafka security configuration.  These properties must be present so that
 # clowder-quarkus-config-source will populate them from the Clowder provided configuration JSON.
@@ -148,7 +157,6 @@ quarkus.smallrye-openapi.path=/api/${quarkus.application.name}/internal/openapi
 # By default, the openapi and swagger ui endpoints are exposed on the management port, so we need to disable it
 # to expose it on the server port 8000:
 quarkus.smallrye-openapi.management.enabled=false
-quarkus.smallrye-openapi.sdk.disabled=${DISABLE_OTEL}
 quarkus.swagger-ui.always-include=true
 quarkus.swagger-ui.path=/api/${quarkus.application.name}/internal/swagger-ui
